# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "X-NBA Documentation",
        "description": "This is the API documentation for the X-NBA API.",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/match-teams": {
            "get": {
                "tags": [
                    "Match-teams"
                ],
                "summary": "Get all match teams",
                "operationId": "c2cfbbca6ee20592ddd43a2b1a53d498",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "string",
                                                "example": "1"
                                            },
                                            "role": {
                                                "type": "string",
                                                "example": "home"
                                            },
                                            "score": {
                                                "type": "number",
                                                "example": 110
                                            },
                                            "winner": {
                                                "type": "boolean",
                                                "example": true
                                            },
                                            "match": {
                                                "properties": {
                                                    "id": {
                                                        "type": "string",
                                                        "example": "1"
                                                    },
                                                    "date": {
                                                        "type": "string",
                                                        "example": "2023-01-01T00:00:00.000000Z"
                                                    },
                                                    "location": {
                                                        "type": "string",
                                                        "example": "Boston, MA"
                                                    },
                                                    "stadium": {
                                                        "type": "string",
                                                        "example": "TD Garden"
                                                    },
                                                    "league": {
                                                        "type": "string",
                                                        "example": "NBA"
                                                    }
                                                },
                                                "type": "object"
                                            },
                                            "team": {
                                                "properties": {
                                                    "id": {
                                                        "type": "string",
                                                        "example": "1"
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Los Angeles Lakers"
                                                    },
                                                    "slug": {
                                                        "type": "string",
                                                        "example": "los-angeles-lakers"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Match-teams"
                ],
                "summary": "Create a match team",
                "operationId": "1b8b28f32396b7b71a26eb18113a7cf2",
                "requestBody": {
                    "description": "Match team data",
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "match_id": {
                                        "type": "string",
                                        "example": "1"
                                    },
                                    "team_id": {
                                        "type": "string",
                                        "example": "1"
                                    },
                                    "role": {
                                        "type": "string",
                                        "example": "home"
                                    },
                                    "score": {
                                        "type": "number",
                                        "example": 110
                                    },
                                    "winner": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "string",
                                            "example": "1"
                                        },
                                        "role": {
                                            "type": "string",
                                            "example": "home"
                                        },
                                        "score": {
                                            "type": "number",
                                            "example": 110
                                        },
                                        "winner": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "match": {
                                            "properties": {
                                                "id": {
                                                    "type": "string",
                                                    "example": "1"
                                                },
                                                "date": {
                                                    "type": "string",
                                                    "example": "2023-01-01T00:00:00.000000Z"
                                                },
                                                "location": {
                                                    "type": "string",
                                                    "example": "Boston, MA"
                                                },
                                                "stadium": {
                                                    "type": "string",
                                                    "example": "TD Garden"
                                                },
                                                "league": {
                                                    "type": "string",
                                                    "example": "NBA"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "team": {
                                            "properties": {
                                                "id": {
                                                    "type": "string",
                                                    "example": "1"
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "Los Angeles Lakers"
                                                },
                                                "slug": {
                                                    "type": "string",
                                                    "example": "los-angeles-lakers"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The match_id field is required."
                                        },
                                        "errors": {
                                            "properties": {
                                                "name": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "The match_id field is required."
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}